pyproject.toml
[project]
name = "rss"
version = "0.1.0"
description = "RSS international news aggregator with DingTalk push"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0",
    "httpx>=0.28",
    "google-genai>=1.0",
    "apscheduler>=3.10,<4",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "openpyxl>=3.1.5",
]

